requirements.txt

Flask==2.0.1
scikit-learn==1.0.2
numpy==1.22.3
pandas==1.4.2
requests==2.27.1
tld==0.12.6
python-whois==0.7.3
beautifulsoup4==4.10.0
```